# docs/specs/api_contract.yaml
openapi: 3.1.0
info:
  title: Derived API Contract
  version: 1.0.0
paths:
...
  /pickup/schedule:
    post:
      operationId: post_pickup_schedule
      responses:
        '201':
          description: Derived from SPEC.md
          content:
            application/json:
              schema:
                type: object
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
...
  /pickup/{return_id}/reschedule:
    put:
      operationId: put_pickup_return_id_reschedule
      responses:
        '200':
          description: Derived from SPEC.md
          content:
            application/json:
              schema:
                type: object
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
...